Factor label update/create into new component
